% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andersen.R
\name{andersen}
\alias{andersen}
\title{Andersen Selectivity Function}
\usage{
andersen(x, p1, p3, p4, p5)
}
\arguments{
\item{x}{Points where Andersen function (see Details) must be computed.}

\item{p1}{Define the value of \eqn{x} at which the transition between the two gaussian functions happens; that is, \eqn{x=p5/exp(p1)}. Negative values are NOT valid.}

\item{p3}{Ascending slope parameter. Negative values or zero are NOT valid.}

\item{p4}{Descending slope parameter.  Negative values or zero are NOT valid.}

\item{p5}{Define the value of \eqn{x} at which the transition between the two gaussian functions happens; that is, \eqn{x=p5/exp(p1)}.  Negative values or zero are NOT valid.}
}
\value{
{SA: }{Vector containing the values of Andersen function in each of the x-values.}
}
\description{
Computes Andersen selectivity function.
}
\details{
The function implements Andersen selectivity function whose equation is \deqn{S(x)<-p0+p2exp(-(ln(p5/x)-p1)^2/p4) if ln(p5/x)<=p1} and \deqn{S(x)<-p0+p2exp(-(ln(p5/x)-p1)^2/p3) if ln(p5/x)>p1.} We fixed \eqn{p0=0}, which is the beginning size for the plateau, and \eqn{p2=1}, which is the maximum value attainable by the function. Remember that \eqn{p3} and \eqn{p4} are the ascending and descending slope parameters, respectively, whereas \eqn{p1} and \eqn{p5} define the value of \eqn{x} at which the transition between the two gaussian functions happens; that is, \eqn{x=p5/exp(p1)}.
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
}
}
