% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BYR_eq.R
\name{BYR.eq}
\alias{BYR.eq}
\title{Total Yield, Biomass and Recruitment in Equilibrium}
\usage{
BYR.eq(Pop.Mod, f.grid, Fish.years, Bio.years, plot, Method, par)
}
\arguments{
\item{Pop.Mod}{A list containing the components returned by Population.Modeling function (main function).}

\item{f.grid}{A sequence of fishing efforts.}

\item{Fish.years}{The number of recent years to estimate the mean of SEL (selectivity, see information about such element in Sum.Pop.Mod function).}

\item{Bio.years}{The number of recent years to estimate the mean of M, Mat, WC, and W (natural mortality, maturity, stock weight and catches weight, see information about such elements in Population.Modeling function and Sum.Pop.Mod function).}

\item{plot}{A vector of two elements. The first one is a logical parameter. By default is equal to TRUE, which means that a biomass per recruit and yield per recruit graphs are done. The second element refers to which iteration must be plotted.}

\item{Method}{The procedure to obtain the age vector of weight (stock and catches), natural mortality, selectivity and maturity. By default is "mean" which means that the mean of the last "Bio.years" is used. The alternative option is "own", the user can introduce these elements.}

\item{par}{If Method="own" it is a list containing the matrices whose columns report for each iteration the age vector of weight (stock and catches), natural mortality, selectivity and maturity. In other case is equal to NULL.}
}
\value{
A list:
\item{N:}{An array whose third dimension is the number of iterations. For each iteration it reports a matrix containing the population size in equilibrium for each age.}
\item{DPM:}{An array whose third dimension is the number of iterations. For each iteration it reports a matrix containing the total biomass in equilibrium, total yield in equilibrium, total recruitment in equilibrium, biomass-per-recruit, and yield-per-recruit for a range of overall fishing mortalities.}
}
\description{
Returns total Yield, Biomass and Recruiment in equilibrium for the corresponding fishing efforts and fishing mortalies. Furthermore, for each case the function also returns the corresponding biomass-per-recruit and yield-per-recruit.
}
\details{
The function returns total Yield, Biomass and Recruiment in equilibrium for the corresponding fishing efforts and fishing mortalities. Furthermore, for each case the function also returns the corresponding biomass-per-recruit and yield-per-recruit. Furthermore, the function also reports the population size in equilibrium for each age and fishing mortality.
}
\examples{
ctrPop<-list(years=seq(1980,2020,by=1),niter=2,N0=15000,ages=0:15,minFage=2,
             maxFage=5,tc=0.5,seed=NULL)
number_ages<-length(ctrPop$ages);number_years<-length(ctrPop$years)
Mvec=c(1,0.6,0.5,0.4,0.35,0.35,0.3,rep(0.3,9))
M<-matrix(rep(Mvec,number_years),ncol = number_years)
colnames(M)<-ctrPop$years
rownames(M)<-ctrPop$ages
ctrBio<-list(M=M,CV_M=0.2, L_inf=20, t0=-0.25, k=0.3, CV_L=0, CV_LC=0, a=6*10^(-6), b=3,
           a50_Mat=1, ad_Mat=-0.5,CV_Mat=0)
ctrSEL<-list(type="cte", par=list(cte=0.5),CV_SEL=0)
f=matrix(rep(0.5,number_years),ncol=number_years,nrow=2,byrow=TRUE)
ctrFish<-list(f=f,ctrSEL=ctrSEL)
a_BH=15000; b_BH=50; CV_REC_BH=0
SR<-list(type="BH",par=c(a_BH,b_BH,CV_REC_BH))
Pop.Mod<-Population.Modeling(ctrPop=ctrPop,ctrBio=ctrBio,ctrFish=ctrFish,SR=SR)
f.grid<-seq(0.00,0.5,by=0.01)
RE<-BYR.eq(Pop.Mod,f.grid,3,3,c(TRUE,1),Method="mean",par=NULL)
N_eq<-RE$N
DPM<-RE$DPM
# The following commented lines refers to par argument.
# If par is not NULL must be something like (assuming that W, WC, M,
# Mat and SEL are defined previously).
# par=list(); par$W<-W; par$SEL<-SEL; par$Mat<-Mat; par$M<-M; par$WC<-WC
# RE=BYR.eq(Pop.Mod,f.grid,plot=c(TRUE,1),Method="own",par=par)
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
\item{Maria Grazia Pennino}
}
}
