% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock_from_Rfishpop.R
\name{FLStock.from.Rfishpop}
\alias{FLStock.from.Rfishpop}
\title{A FLStock class object}
\usage{
FLStock.from.Rfishpop(Pop.Mod)
}
\arguments{
\item{Pop.Mod}{A list containing the components returned by Population.Modeling function (main function).}
}
\value{
{The FLStock object (see R package FLCore) corresponding to the simulated population generated by Population.Modeling function.}
}
\description{
Defines and fill a FLStock object using the information provided by the main function Population.Modeling.
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
}
}
