% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma.R
\name{gamma_SEL}
\alias{gamma_SEL}
\title{Gamma Selectivity Function}
\usage{
gamma_SEL(x, alpha, beta, gamma)
}
\arguments{
\item{x}{Points where Gamma function must be computed (see Details).}

\item{alpha}{the shape parameter. Negative values are NOT valid.}

\item{beta}{the scale parameter. Negative values or zero are NOT valid.}

\item{gamma}{the size of the mesh used by the fleet (according to the units of the remaining parameters, for example, cm). Negative values or zero are NOT valid.}
}
\value{
{SG: }{Vector containing the values of Gamma function in each of the x-values.}
}
\description{
Computes Gamma selectivity function.
}
\details{
The function implements Gamma selectivity function which equation is \deqn{S(x)<-((x/((alpha-1) beta gamma))^(alpha-1))exp(alpha-1-(1/beta gamma)),} where gamma is the size of the mesh, alpha is the shape parameter and beta is the scale parameter.
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
}
}
