% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic.R
\name{Logistic}
\alias{Logistic}
\title{Logistic function}
\usage{
Logistic(x, x50, xd)
}
\arguments{
\item{x}{Points where the logistic function must be computed.}

\item{x50}{x-value of the sigmoid's midpoint of the logistic function.}

\item{xd}{Minus the inverse of the logistic growth rate (steepness of the curve).}
}
\value{
{Vector containing the values of the logistic function in each of the x-values.}
}
\description{
Computes the logistic function.
}
\details{
The function implements the logistic function which equation is \deqn{L(x)=1/(1+exp((x-x50)/xd))} where x50 is the x-value of the sigmoid's midpoint of the logistic function, and xd is minus the inverse of the logistic growth rate (steepness of the curve).
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
}
}
