% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruitment_B.R
\name{RBH}
\alias{RBH}
\title{Beverton-Holt Recruitment Model}
\usage{
RBH(SSB, a, b)
}
\arguments{
\item{SSB}{Maturity biomass for each year (spawning stock)}

\item{a}{Maximum number of recruitments produced.}

\item{b}{Spawning stock needed to produce recruitment equal to half maximum.}
}
\value{
{Vector containing the recruitment for each year.}
}
\description{
Computes the recruitment using Beverton-Holt Recruitment Model.
}
\details{
The formulation of the Beverton-Holt Recruitment Model equation is: \deqn{R=a*SSB/(b+SSB)} where SSB is the Maturity biomass (spawning stock), a is the maximum number of recruitments produced and b is the spawning stock needed to produce recruitment equal to half maximum.
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
}
}
