% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recruitment_R.R
\name{RRK}
\alias{RRK}
\title{Ricker Recruitment Model}
\usage{
RRK(SSB, a, b)
}
\arguments{
\item{SSB}{Maturity biomass for each year (spawning stock)}

\item{a}{Recruits-per-spawner at low stock levels.}

\item{b}{Relates to the rate of decreasing of recruits-per-spawner as SSB increases.}
}
\value{
{Vector containing the recruitment for each year.}
}
\description{
Computes the recruitment using Ricker Model.
}
\details{
The formulation of the Ricker equation is: \deqn{R=a*SSB*exp(-b*SSB)} where SSB is the Maturity biomass (spawning stock), a is the recruits-per-spawner at low stock levels and b relates to the rate of decreasing of recruits-per-spawner as SSB increases.
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
}
}
