% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecting_units.R
\name{selecting_units}
\alias{selecting_units}
\title{Selecting Units}
\usage{
selecting_units(Pop.Mod, Elements)
}
\arguments{
\item{Pop.Mod}{A list containing the components returned by Population.Modeling function (main function).}

\item{Elements}{A vector specifing which of the following elements must be reported by the function.\itemize{
\item{"C":}{Weight of the captures for each year and iteration.}
\item{"BIO":}{Total biomass for each year and iteration.}
\item{"SSB":}{Maturity biomass for each year (spawning stock) and iteration.}}}
}
\value{
A list containing the objects specified before using argument "Elements" in tonnes.
}
\description{
Returns some of the matrices provided by Sum.Pop.Mod function but in tonnes instead of Kg.
}
\examples{

#selecting_units(Pop.Mod,c("C","BIO","SSB"))
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
}
}
