% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight.R
\name{Weight}
\alias{Weight}
\title{Length-Weight relationship}
\usage{
Weight(L, a, b)
}
\arguments{
\item{L}{Vector of lengths by age.}

\item{a}{Allometric growth parameter.}

\item{b}{Scaling constant.}
}
\value{
{Vector containing the weight for each age.}
}
\description{
Computes the weight from the length using the relationship explained in Details.
}
\details{
The function implements the following Length-Weight relationship
 \deqn{W=a*L^b} where L is the vector of lengths by age, a is the allometric growth parameter, b scaling constant, and W is the age weight vector.
}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
}
}
