% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF.R
\name{RF}
\alias{RF}
\title{Reference Points}
\usage{
RF(
  Pop.Mod,
  Fish.years,
  Bio.years,
  Method,
  par,
  FM_type,
  iters,
  plot,
  prop = 0.3
)
}
\arguments{
\item{Pop.Mod}{A list containing the components returned by Population.Modeling function (main function).}

\item{Fish.years}{The number of recent years to estimate the mean of SEL (selectivity).}

\item{Bio.years}{The number of recent years to estimate the mean of M, Mat, WC, and W (natural mortality, maturity, stock weight and catch weight).}

\item{Method}{The procedure to obtain the age vector of weight (stock and catches), natural mortality, selectivity and maturity. By default is "mean" which means that the mean of the last "Bio.years" is used. The alternative option is "own", the user can introduce these elements.}

\item{par}{If Method="own" it is a list containing the age vector of weight (stock and catches), natural mortality, selectivity and maturity (for the first iteration). In other case is equal to NULL.}

\item{FM_type}{a vector containing which of the five reference fishery mortalities must be computed. The possibilities have been described above: FM_type="F_max", FM_type="F_0.1", FM_type="F_msy", FM_type="F_Crash" and FM_type="F_BPR".}

\item{iters}{A vector containing the iterations for which the reference fishery mortalities must be computed.}

\item{plot}{A logical parameter equal to TRUE if the user desires to obtain several informative plots (see details).}

\item{prop}{The proportion of BPR at virgin biomass at which FM_type="F_BPR" refers. By default is 0.3. If the user desires the result for different proportions a vector containing the different values must be introduced using prop argument.}
}
\value{
The function returns a list reporting arrays (third dimension *iters*) containing the corresponding elements of the following list depending the above selection (*FM_type* argument). \itemize{
\item{F_max:}{the value of F that produces maximum YPR with the corresponding effort of fishing, YPR, BYR, B, Y and R in equilibrium.}
\item{F_0.1:}{the value of F at which the slope of the YPR curve is reduced to 0.1 of that estimated at the origin with the corresponding effort of fishing, YPR, BYR, B, Y and R in equilibrium.}
\item{F_msy:}{the value of F at which the MSY is attained with the corresponding effort of fishing, YPR, BYR, B, Y and R in equilibrium.}
\item{F_Crash:}{the value of F which will drive the stock to extinction with the corresponding effort of fishing, YPR, BYR, B, Y and R in equilibrium.}
\item{F_BPR:}{the value of F at which at which the BPR is equal to prop multiplied by BPR0, where prob is any proportion that the user can fix and BPR0 is the BPR at virgin biomass, with the corresponding effort of fishing, YPR, BYR, B, Y and R in equilibrium. To access to this element use F_BPRprop following by symbol of percentage (without spaces and between quotation marks) where prop is the value that the user specifies. If prop vector length is greater than 1 the function will return several slots in the list reporting the corresponding values for each value of prop. The user can access to the elements as we mentioned before replacing prop for any of the value in such vector.}
Futhermore the function returns also an slot termed INFO containing some information useful for the application of other functions of the package.

}
}
\description{
Returns the reference fishery mortality which produces maximum YPR (FM_type="F_max"), the reference fishery mortality at which the slope of the YPR curve is reduced to 0.1 of that estimated at the origin (FM_type="F_0.1"), the reference fishery mortality at which the MSY is attained (FM_type="F_msy"), the reference fishery mortality which will drive the stock to extinction (FM_type="F_Crash") and the reference fishery mortality at which the BPR is equal to prop multiplied by BPR0 where prop is any proportion that the user can fix and BPR0 is the BPR at virgin biomass (FM_type="F_BPR"). Furthermore for each of these fishery mortalities the corresponding effort of fishing, YPR, BYR, B, Y and R in equilibrium is also returned.
}
\details{
The function returns the reference fishery mortality which produces maximum YPR (FM_type="F_max"), the reference fishery mortality at which the slope of the YPR curve is reduced to 0.1 of that estimated at the origin (FM_type="F_0.1"), the reference fishery mortality at which the MSY is attained (FM_type="F_msy"), the reference fishery mortality which will drive the stock to extinction (FM_type="F_Crash") and and the reference fishery mortality at which the BPR is equal to prop multiplied by BPR0 where prop is any proportion that the user can fix and BPR0 is the BPR at virgin biomass (FM_type="F_BPR"). Furthermore for each of these fishery mortalities the corresponding effort of fishing, YPR, BYR, B, Y and R in equilibrium is also returned. If the fishing effort is equal to 10 can be that optimize process had not found the correct value in the default sequence (except for FM_type="F_Crash", in this case the sequence finishes at 60).

If plot=TRUE, the function reports the plots: equilibrium biomass v. fishing mortality, equilibrium yield v. fishing mortality, equilibrium recruitment v. biomass,  and equilibrium yield v. biomass. All the plots are done using the infromation of the first iteration in the vector of parameter *iters*. The corresponding reference fishery mortalities are also plotted in the previous curves. Note that only the fishery mortalities required by the argument FM_type are plotted. If prop is a vector of length greater than 1 only the values corresponding to the first element of the vector will be represented in the plot for simplicity. On the other hand, note that F_msy and F_max coincide when the recruitment relationship is constant and hence only one of both appears in the plot. The reference fishery mortalities F_max and F_Crash can be overlapped in equilibrium recruitment v. biomass and equilibrium yield v. biomass plots then only one can be shown in the plot.

Note that F_Crash does not exist when the recruitment model is constant.
}
\examples{
ctrPop<-list(years=seq(1980,2020,by=1),niter=2,N0=15000,ages=0:15,minFage=2,
             maxFage=5,tc=0.5,seed=NULL)
number_ages<-length(ctrPop$ages);number_years<-length(ctrPop$years)
Mvec=c(1,0.6,0.5,0.4,0.35,0.35,0.3,rep(0.3,9))
M<-matrix(rep(Mvec,number_years),ncol = number_years)
colnames(M)<-ctrPop$years
rownames(M)<-ctrPop$ages
ctrBio<-list(M=M,CV_M=0.2, L_inf=20, t0=-0.25, k=0.3, CV_L=0, CV_LC=0, a=6*10^(-6), b=3,
           a50_Mat=1, ad_Mat=-0.5,CV_Mat=0)
ctrSEL<-list(type="cte", par=list(cte=0.5),CV_SEL=0)
f=matrix(rep(0.5,number_years),ncol=number_years,nrow=2,byrow=TRUE)
ctrFish<-list(f=f,ctrSEL=ctrSEL)
a_BH=15000; b_BH=50; CV_REC_BH=0
SR<-list(type="BH",par=c(a_BH,b_BH,CV_REC_BH))
Pop.Mod<-Population.Modeling(ctrPop=ctrPop,ctrBio=ctrBio,ctrFish=ctrFish,SR=SR)
# Uncomment the following lines:
# We compute the reference fishery mortalities.
# RF(Pop.Mod, 3,3,Method="mean",par=NULL,FM_type="F_max",iters=1:2,plot=TRUE)
# RF(Pop.Mod, 3,3,Method="mean",par=NULL,FM_type="F_0.1",iters=1:2,plot=TRUE)
# RF(Pop.Mod, 3,3,Method="mean",par=NULL,FM_type="F_msy",iters=1:2,plot=TRUE)
# RF(Pop.Mod, 3,3,Method="mean",par=NULL,FM_type="F_Crash",iters=1:2,plot=TRUE)
# RF(Pop.Mod, 3,3,Method="mean",par=NULL,FM_type="F_BPR",iters=1:2,plot=TRUE,prop=0.4)
# RF(Pop.Mod, 3,3,Method="mean",par=NULL,FM_type=c("F_Crash","F_msy"),iters=1:2
#,plot=TRUE)
# RF(Pop.Mod, 3,3,Method="mean",par=NULL,FM_type=c("F_Crash","F_msy",
#"F_0.1","F_max","F_BPR")
#,iters=1:2,plot=TRUE)

# The following commented lines refers to par argument.
# If par is not NULL must be something like (assuming that W, WC, M,
# Mat and SEL are defined previously).
# par=list(); par$W<-W; par$SEL<-SEL; par$Mat<-Mat; par$M<-M; par$WC<-WC
# RF(Pop.Mod,Method="own",par=par2,FM_type=c("F_Crash","F_msy",
#"F_0.1","F_max","F_BPR"),iters=1:2,plot=TRUE)

}
\author{
\itemize{
\item{Marta Cousido-Rocha}
\item{Santiago Cerviño López}
\item{Maria Grazia Pennino}
}
}
